Separate sessionserver from server_controller.py
Separate sessionserver from server_controller.py
